feature: improve validation (#3)
* feat: adds CEP validation method

Adds a method to validate CEP using an external API.

This new method enables the application to verify the validity of CEPs entered by users, improving data accuracy and providing a better user experience.

* fix: Updates backend dependency and documentation

* feat: implements integrated person search

Adds a search component for integrated persons.

This component allows users to search for integrated persons
by CPF and displays their details, including name,
birthdate, status, inclusion date/time and last update
date/time.

The component uses a mock data object as a placeholder
before retrieving data from a backend.

Includes input validation for CPF to limit the input to 11
digits.

* feat: enables pessoa integration

Adds functionality to trigger pessoa integration from the list view.
The "Integrar" button is now functional, calling the provided `onIntegrar` prop.
It's conditionally rendered based on the pessoa's integration status.

* feat: implements pessoa integration and deletion

Adds functionality to integrate and delete 'pessoa' entities,
enhancing data management capabilities.

Introduces a search feature for integrated 'pessoa' entities
based on CPF, improving data retrieval efficiency.

Updates the UI to reflect these new features, providing a
more comprehensive user experience.

* feat: implements CEP validation and address fill

This commit introduces CEP validation and automatic address filling functionality to the cadastro page.

It adds a new state variable to manage the visibility of the CEP search icon and whether the address fields should be auto-filled and read-only based on CEP lookup.

Includes validation for CEP format and mandatory address fields when CEP lookup fails.

The cepCodeValidation function fetches address data from an external API using the entered CEP, populating the corresponding address fields if successful. If the API returns an error or the CEP is not found, appropriate error messages are logged, and the address fields are cleared.

* feat: add extra validation to Pessoa.nome

- Check if nome has 2 words or more.
- Check if nome has uppercase letter for each word

* feat: implements cadastro validations

This commit implements validations for the cadastro of a pessoa.
It checks if the birth date is in the future and if the CPF has 11 digits.
Also sends pessoa for validation if all information is complete.